refactor: delete method with dialog component

+++ crud-angular/src/app/courses/containers/courses/courses.component.ts
@@ -9,6 +9,9 @@ import {ErrorDialogComponent} from "../../../shared/components/error-dialog/erro
 import {ActivatedRoute, Router} from "@angular/router";
 import {resetCumulativeDurations} from "@angular-devkit/build-angular/src/builders/browser-esbuild/profiling";
 import {MatSnackBar} from "@angular/material/snack-bar";
+import {
+  ConfirmationDialogComponent
+} from "../../../shared/components/confirmation-dialog/confirmation-dialog.component";
 
 @Component({
   selector: 'app-courses',
@@ -56,16 +59,25 @@ onEdit(course: Course) {
 }
 
   onRemove(course: Course) {
-    this.coursesService.remove(course._id).subscribe(
-      () => {
-        this.refresh();
-        this.snackBar.open('Curso removido com sucesso!', 'X', {
-          duration: 5000,
-          verticalPosition: 'top',
-          horizontalPosition: 'center'
-        });
-      },
-      () => this.onError('Erro ao tentar remover curso.')
-    );
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja remover esse curso?',
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.coursesService.remove(course._id).subscribe(
+          () => {
+            this.refresh();
+            this.snackBar.open('Curso removido com sucesso!', 'X', {
+              duration: 5000,
+              verticalPosition: 'top',
+              horizontalPosition: 'center'
+            });
+          },
+          () => this.onError('Erro ao tentar remover curso.')
+        );
+      }
+    });
   }
+
 }
